Add meta scraper Booking class
Responsible to arbitrate between Reservation + Hotel scraping, or just Hotel scraping, based on the housing url provided by the user
