+++ apps/web/templates/trips/housings/new.html.erb
@@ -35,7 +35,7 @@
 <div class="booking-container">
   <div class="booking-notice">
     <div class="title">
-      <span class="icon">⚠</span>
+      <span class="icon">⚠️</span>
     <h2>
       Adding a housing from Booking.com
     </h2>
