+++ lib/housing_list/interactors/housings/fetch_details_and_update.rb
@@ -15,7 +15,7 @@ module Housings
 
     def call
       provider_name       = housing.provider.split('_').map(&:capitalize).join
-      scraper_class_name  = "Scrapers::#{provider_name}"
+      scraper_class_name  = "Scrapers::#{provider_name}::Housing"
       scraper_class       = Object.const_get(scraper_class_name)
       housing_attributes  = scraper_class.new(url: housing.url).scrape
